Too many things to list
Don't remember everything that changed, but I did make a lot of text dependent on the lang file so they can be translated now, had the benefit of significantly reducing the amount of giant lengths of if statements in some of the classes.
Additionally, I tried to reduce that sort of thing in the starter kit and grenade classes. The starter kit now receives arrays in its constructor which it uses for the giving to inventory, works most of the time, but a few that used [Arrays.fill] are buggy, working on it. The grenade class also now can take the grenade entity's class as a parameter in the constructor as well, surprisingly bugless.
Oh right, energy guns and a bunch of things that are references, we'll see how they develop.
Remember when this fork was just about supercomputers and storage devices?
Oh, so I did end up listing them...