Various additions and reworks
A lot of changes, don't know what else to say. Summary-ish:

I will say that I added a lot of heavier transuranics (not all have to be added, just to see if they'll fit in) as well as U-234 which comes from the decay of Pu-238 RTGs. Also plan on it to come from spent RBMK fuel, but alas I don't have that in my fork so I can't add it.
Changed the custom toxicity handler to use an enumerate since that makes more sense.
Added a "chad" screwdriver since main has the screwdriver break.
Misc. changes to various classes to reduce spaghetti slightly, will probably never be merged but it makes me happier.
Added rounding to the tooltips of radioactive things, useful for when realistic RAD/s is on.
Plans on Cf-252 + Zirconium neutron sources, for the RBMK of course, extremely strong. Maybe I could add it to the breeding reactor's fuel?
Possibly more things, but I don't remember them all!