Revamped OpenComputers compatibility, the following is a list of changes/additions.
Added compatibility to the large launch pad
Changed the way that fluids are handled by OC, they now use the internal names of said fluids.
Changing over to a more compact function for energy related compatibility, not doing it fully until people have a bit of time to modify their programs.
Fixed a minor bug where changing the steam type using OC in a RBMK boiler could potentially make infinite steam.
Wiki will be updated for the compatibility on merge of this PR.