Erythraean Esprit Kouprey
More fixes for more bugs (that I introduced), mostly centered around reloading and trying to work with guns that don't use the magazine system while the system is active.
Also features a fix for the Benelli M4's ejector definition. The y-offset was 1 and z-offset 0 for some reason, probably from the previous issue in my testing.